Fix other tests which relied on old init logic
